Fully split CustomNodeView into a separate, memoized component
Plain custom node views had been broken since we implemented SSR.
When they were pulled into their own component, it messed up some of
the layout effect ordering and triggering, such that we no longer
attached them to the DOM.

To resolve this, the NodeView component now either renders ReactNodeView
or CustomNodeView, each of which manages its own full lifecycle, decorations,
etc.
