+++ .yarn/versions/1cfdbe48.yml
@@ -1,0 +1,2 @@
+releases:
+  "@handlewithcare/react-prosemirror": patch
